Projects/NLP Apps/Aman_Singh/requirements.txt
python-dotenv
langchain
langchain_groq
langchain_community
langchain_experimental
langchain_core
streamlit
tabulate
